refactor: remove dead WS protocol types, split fs types into transport::ws
